Move NodeCmd related module out of EraBased folder
