+++ cardano-cli/src/Cardano/CLI/Commands.hs
@@ -8,9 +8,9 @@ where
 import           Cardano.CLI.Byron.Commands (ByronCommand)
 import           Cardano.CLI.Commands.Debug
 import           Cardano.CLI.Commands.Hash (HashCmds)
+import           Cardano.CLI.Commands.Node
 import           Cardano.CLI.Commands.Ping (PingCmd (..))
 import           Cardano.CLI.EraBased.Commands
-import           Cardano.CLI.EraBased.Commands.Node
 import           Cardano.CLI.Legacy.Commands
 
 import           Options.Applicative.Types (ParserInfo (..), ParserPrefs (..))
